+++ src/main/java/eu/gir/girsignals/blocks/GhostBlock.java
@@ -28,29 +28,28 @@ public class GhostBlock extends Block {
 	}
 
 	@Override
-    public EnumBlockRenderType getRenderType(IBlockState state)
-    {
-        return EnumBlockRenderType.INVISIBLE;
-    }
+	public EnumBlockRenderType getRenderType(IBlockState state) {
+		return EnumBlockRenderType.INVISIBLE;
+	}
 
-    @SideOnly(Side.CLIENT)
+	@SideOnly(Side.CLIENT)
 	@Override
-    public float getAmbientOcclusionLightValue(IBlockState state)
-    {
-        return 1.0F;
-    }
+	public float getAmbientOcclusionLightValue(IBlockState state) {
+		return 1.0F;
+	}
 
 	@Override
-    public void dropBlockAsItemWithChance(World worldIn, BlockPos pos, IBlockState state, float chance, int fortune)
-    {
-    }
-	
+	public void dropBlockAsItemWithChance(World worldIn, BlockPos pos, IBlockState state, float chance, int fortune) {
+	}
+
 	@Override
 	public ItemStack getPickBlock(IBlockState state, RayTraceResult target, World world, BlockPos pos,
 			EntityPlayer player) {
-		return SignalBlock.pickBlock(player);
+		BlockPos downPos = pos.down();
+		Block lowerBlock = world.getBlockState(downPos).getBlock();
+		return lowerBlock.getPickBlock(state, target, world, downPos, player);
 	}
-	
+
 	@Override
 	public boolean isOpaqueCube(IBlockState state) {
 		return false;
@@ -88,7 +87,7 @@ public class GhostBlock extends Block {
 
 		BlockPos posdown = pos.down();
 		Block lowerBlock = worldIn.getBlockState(posdown).getBlock();
-		if (lowerBlock instanceof GhostBlock || lowerBlock instanceof SignalBlock) {
+		if (lowerBlock instanceof GhostBlock || lowerBlock instanceof Signal) {
 			worldIn.destroyBlock(posdown, false);
 		}
 	}