+++ src/main/java/eu/gir/girsignals/blocks/SignalEL.java
@@ -4,11 +4,12 @@ import eu.gir.girsignals.EnumSignals.EL;
 import eu.gir.girsignals.EnumSignals.EL_ARROW;
 import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.SEProperty.ChangeableStage;
+import eu.gir.girsignals.init.GIRItems;
 
-public class SignalEL extends SignalBlock {
+public class SignalEL extends Signal {
 
 	public SignalEL() {
-		super("elsignal", 2);
+		super(GIRItems.SIGN_PLACEMENT_TOOL, "elsignal", 2);
 	}
 
 	public static final SEProperty<EL> ELTYPE = SEProperty.of("eltype", EL.EL1V, ChangeableStage.GUISTAGE);