+++ src/main/java/eu/gir/girsignals/blocks/signals/SignalKS.java
@@ -1,16 +1,18 @@
-package eu.gir.girsignals.blocks;
+package eu.gir.girsignals.blocks.signals;
 
-import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.EnumSignals.KS;
 import eu.gir.girsignals.EnumSignals.KS_DISTANT;
 import eu.gir.girsignals.EnumSignals.MAST_SIGN;
 import eu.gir.girsignals.EnumSignals.ZS32;
+import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.SEProperty.ChangeableStage;
+import eu.gir.girsignals.blocks.Signal;
+import eu.gir.girsignals.init.GIRItems;
 
-public class SignalKS extends SignalBlock {
+public class SignalKS extends Signal {
 
 	public SignalKS() {
-		super("KS", 6, 4.95f);
+		super(GIRItems.PLACEMENT_TOOL, "KS", 6, 4.95f);
 	}
 
 	public static final SEProperty<KS> STOPSIGNAL = SEProperty.of("kombisignal", KS.OFF);