+++ src/main/java/eu/gir/girsignals/blocks/signals/SignalLF.java
@@ -1,14 +1,16 @@
-package eu.gir.girsignals.blocks;
+package eu.gir.girsignals.blocks.signals;
 
 import eu.gir.girsignals.EnumSignals.LF;
 import eu.gir.girsignals.EnumSignals.LFBACKGROUND;
 import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.SEProperty.ChangeableStage;
+import eu.gir.girsignals.blocks.Signal;
+import eu.gir.girsignals.init.GIRItems;
 
-public class SignalLF extends SignalBlock {
+public class SignalLF extends Signal {
 
 	public SignalLF() {
-		super("lfsignal", 1);
+		super(GIRItems.SIGN_PLACEMENT_TOOL, "lfsignal", 1);
 	}
 
 	public static final SEProperty<LF> INDICATOR = SEProperty.of("indicator", LF.Z1, ChangeableStage.GUISTAGE);