+++ src/main/java/eu/gir/girsignals/blocks/signals/SignalTram.java
@@ -1,14 +1,16 @@
-package eu.gir.girsignals.blocks;
+package eu.gir.girsignals.blocks.signals;
 
-import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.EnumSignals.CAR;
 import eu.gir.girsignals.EnumSignals.PED;
 import eu.gir.girsignals.EnumSignals.TRAM;
+import eu.gir.girsignals.SEProperty;
+import eu.gir.girsignals.blocks.Signal;
+import eu.gir.girsignals.init.GIRItems;
 
-public class SignalTram extends SignalBlock {
+public class SignalTram extends Signal {
 
 	public SignalTram() {
-		super("TramSignal", 0);
+		super(GIRItems.PLACEMENT_TOOL, "TramSignal", 0);
 	}
 
 	public static final SEProperty<TRAM> TRAMSIGNAL = SEProperty.of("signaltram", TRAM.OFF);