+++ src/main/java/eu/gir/girsignals/items/Placementtool.java
@@ -1,10 +1,12 @@
 package eu.gir.girsignals.items;
 
+import java.util.ArrayList;
+
 import eu.gir.girsignals.EnumSignals.IIntegerable;
 import eu.gir.girsignals.GirsignalsMain;
 import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.SEProperty.ChangeableStage;
-import eu.gir.girsignals.blocks.SignalBlock;
+import eu.gir.girsignals.blocks.Signal;
 import eu.gir.girsignals.guis.GUIHandler;
 import eu.gir.girsignals.init.GIRBlocks;
 import eu.gir.girsignals.init.GIRNetworkHandler;
@@ -21,15 +23,18 @@ import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.world.World;
 import net.minecraftforge.common.property.ExtendedBlockState;
 
-public class Placementtool extends Item implements IIntegerable<SignalBlock> {
+public class Placementtool extends Item implements IIntegerable<Signal> {
 
-	private SignalBlock[] blocks;
-	
-	public Placementtool(SignalBlock... blocks) {
-		this.blocks = blocks;
+	private final ArrayList<Integer> signalids = new ArrayList<>();
+
+	public Placementtool() {
 		setCreativeTab(GIRTabs.tab);
 	}
 
+	public void addSignal(final Signal sig) {
+		signalids.add(sig.getID());
+	}
+	
 	@SuppressWarnings({ "rawtypes", "unchecked" })
 	@Override
 	public EnumActionResult onItemUse(EntityPlayer player, World worldIn, BlockPos pos, EnumHand hand,
@@ -50,7 +55,7 @@ public class Placementtool extends Item implements IIntegerable<SignalBlock> {
 				player.sendMessage(new TextComponentTranslation("pt.itemnotset"));
 				return EnumActionResult.FAIL;
 			}
-			SignalBlock block = SignalBlock.SIGNALLIST.get(compound.getInteger(GIRNetworkHandler.BLOCK_TYPE_ID));
+			Signal block = Signal.SIGNALLIST.get(compound.getInteger(GIRNetworkHandler.BLOCK_TYPE_ID));
 			int height = block.getHeight(compound);
 			BlockPos lastPos = setPosition;
 			for (int i = 0; i < height; i++)
@@ -71,36 +76,28 @@ public class Placementtool extends Item implements IIntegerable<SignalBlock> {
 				SEProperty sep = SEProperty.cst(iup);
 				if (sep.isChangabelAtStage(ChangeableStage.GUISTAGE)) {
 					sig.setProperty(sep, sep.getObjFromID(compound.getInteger(iup.getName())));
-				} else if (sep.isChangabelAtStage(ChangeableStage.APISTAGE) && compound.getBoolean(iup.getName())) {
+				} else if ((sep.isChangabelAtStage(ChangeableStage.APISTAGE) && compound.getBoolean(iup.getName()))
+						|| sep.isChangabelAtStage(ChangeableStage.APISTAGE_NONE_CONFIG)) {
 					sig.setProperty(sep, sep.getDefault());
 				}
 			});
 			String str = compound.getString(GIRNetworkHandler.SIGNAL_CUSTOMNAME);
 			if (!str.isEmpty())
 				sig.setCustomName(str);
+			sig.setBlockID();
 			worldIn.notifyBlockUpdate(setPosition, ebs.getBaseState(), ebs.getBaseState(), 3);
 			return EnumActionResult.SUCCESS;
 		}
 	}
 
 	@Override
-	public SignalBlock getObjFromID(int obj) {
-		return blocks[obj];
-	}
-	
-	public int getTransform(int obj) {
-		int x = 0;
-		for (SignalBlock signalBlock : blocks) {
-			if(signalBlock.getID() == obj)
-				return x;
-			x++;
-		}
-		return x;
+	public Signal getObjFromID(int obj) {
+		return Signal.SIGNALLIST.get(signalids.get(obj));
 	}
 
 	@Override
 	public int count() {
-		return blocks.length;
+		return signalids.size();
 	}
 
 }