+++ src/main/java/eu/gir/girsignals/models/GIRCustomModelLoader.java
@@ -29,19 +29,19 @@ import eu.gir.girsignals.EnumSignals.TRAM;
 import eu.gir.girsignals.EnumSignals.VR;
 import eu.gir.girsignals.EnumSignals.ZS32;
 import eu.gir.girsignals.GirsignalsMain;
+import eu.gir.girsignals.blocks.Signal;
+import eu.gir.girsignals.blocks.Signal.SignalAngel;
 import eu.gir.girsignals.blocks.SignalBUE;
 import eu.gir.girsignals.blocks.SignalBUELight;
-import eu.gir.girsignals.blocks.SignalBlock;
 import eu.gir.girsignals.blocks.SignalEL;
-import eu.gir.girsignals.blocks.SignalBlock.SignalAngel;
-import eu.gir.girsignals.blocks.SignalHL;
-import eu.gir.girsignals.blocks.SignalHV;
-import eu.gir.girsignals.blocks.SignalKS;
-import eu.gir.girsignals.blocks.SignalLF;
 import eu.gir.girsignals.blocks.SignalOTHER;
 import eu.gir.girsignals.blocks.SignalRA;
-import eu.gir.girsignals.blocks.SignalSHLight;
-import eu.gir.girsignals.blocks.SignalTram;
+import eu.gir.girsignals.blocks.signals.SignalHL;
+import eu.gir.girsignals.blocks.signals.SignalHV;
+import eu.gir.girsignals.blocks.signals.SignalKS;
+import eu.gir.girsignals.blocks.signals.SignalLF;
+import eu.gir.girsignals.blocks.signals.SignalSHLight;
+import eu.gir.girsignals.blocks.signals.SignalTram;
 import net.minecraft.client.renderer.block.model.BuiltInModel;
 import net.minecraft.client.renderer.block.model.ItemCameraTransforms;
 import net.minecraft.client.renderer.block.model.ItemOverrideList;
@@ -138,8 +138,8 @@ public class GIRCustomModelLoader implements ICustomModelLoader {
 			cm.register("hv/hv_mast2", ebs -> true, 2);
 			cm.register("hv/hv_mast3", ebs -> true, 3);
 			cm.register("hv/hv_mast4", ebs -> true, 4);
-
-			cm.register("hv/hv_number", has(SignalBlock.CUSTOMNAME), 2);
+			
+			cm.register("hv/hv_number", has(Signal.CUSTOMNAME), 2);
 
 			// Zs1 on
 			cm.register("hv/hv_zs1", hasAndIs(SignalHV.ZS1).and(has(SignalHV.STOPSIGNAL)), 4.4f, "lamp1north",
@@ -205,8 +205,8 @@ public class GIRCustomModelLoader implements ICustomModelLoader {
 			cm.register("ks/ks_sign_distant", has(SignalKS.MASTSIGNDISTANT), 1);
 			cm.register("ks/ks_mast2", ebs -> true, 2);
 
-			cm.register("ks/ks_number", has(SignalBlock.CUSTOMNAME), 4);
-
+			cm.register("ks/ks_number", has(Signal.CUSTOMNAME), 4);
+			
 			for (MAST_SIGN sign : MAST_SIGN.values())
 				if (!sign.equals(MAST_SIGN.OFF))
 					cm.register("ks/ks_sign", with(SignalKS.MASTSIGN, ms -> ms.equals(sign)), 2, "13",
@@ -306,9 +306,9 @@ public class GIRCustomModelLoader implements ICustomModelLoader {
 			cm.register("hl/hl_ne2_2", has(SignalHL.NE2).and(has(SignalHL.NE2_2)).and(has(SignalHL.DISTANTSIGNAL)), 1);
 			cm.register("hl/hl_sign_distant", has(SignalHL.MASTSIGNDISTANT), 1);
 			cm.register("hl/hl_mast2", ebs -> true, 2);
-
-			cm.register("hl/hl_number", has(SignalBlock.CUSTOMNAME), 0);
-
+			
+			cm.register("hl/hl_number", has(Signal.CUSTOMNAME), 0);
+			
 			for (MAST_SIGN sign : MAST_SIGN.values())
 				if (!sign.equals(MAST_SIGN.OFF))
 					cm.register("hl/hl_sign_main", with(SignalHL.MASTSIGN, ms -> ms.equals(sign)), 2, "9",