+++ src/main/java/eu/gir/girsignals/tileentitys/SignalControllerTileEntity.java
@@ -2,30 +2,34 @@ package eu.gir.girsignals.tileentitys;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.BiConsumer;
 
 import eu.gir.girsignals.SEProperty;
 import eu.gir.girsignals.SEProperty.ChangeableStage;
-import eu.gir.girsignals.blocks.SignalBlock;
+import eu.gir.girsignals.blocks.Signal;
 import eu.gir.girsignals.debug.NetworkDebug;
 import eu.gir.girsignals.items.Linkingtool;
 import li.cil.oc.api.machine.Arguments;
 import li.cil.oc.api.machine.Callback;
 import li.cil.oc.api.machine.Context;
 import li.cil.oc.api.network.SimpleComponent;
-import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.multiplayer.ChunkProviderClient;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.network.NetworkManager;
 import net.minecraft.network.play.server.SPacketUpdateTileEntity;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
-import net.minecraftforge.common.property.IExtendedBlockState;
+import net.minecraft.world.chunk.Chunk;
+import net.minecraft.world.chunk.Chunk.EnumCreateEntityType;
+import net.minecraft.world.gen.ChunkProviderServer;
 import net.minecraftforge.fml.common.Optional;
 
 @Optional.Interface(iface = "li.cil.oc.api.network.SimpleComponent", modid = "opencomputers")
 public class SignalControllerTileEntity extends TileEntity implements SimpleComponent {
-
+	
 	private BlockPos linkedSignalPosition = null;
 	private int[] listOfSupportedIndicies;
 	private Map<String, Integer> tableOfSupportedSignalTypes;
@@ -53,7 +57,7 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 	public void readFromNBT(NBTTagCompound compound) {
 		linkedSignalPosition = readBlockPosFromNBT(compound);
 		super.readFromNBT(compound);
-		if(world != null && world.isRemote && linkedSignalPosition != null)
+		if (world != null && world.isRemote && linkedSignalPosition != null)
 			onLink();
 		NetworkDebug.networkReadHook(compound, world, this);
 	}
@@ -65,7 +69,7 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 		NetworkDebug.networkWriteHook(compound, world, this);
 		return compound;
 	}
-	
+
 	@Override
 	public SPacketUpdateTileEntity getUpdatePacket() {
 		return new SPacketUpdateTileEntity(pos, 0, getUpdateTag());
@@ -74,7 +78,7 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 	@Override
 	public void onDataPacket(NetworkManager net, SPacketUpdateTileEntity pkt) {
 		this.readFromNBT(pkt.getNbtCompound());
-		if(hasLinkImpl())
+		if (hasLinkImpl())
 			onLink();
 	}
 
@@ -84,24 +88,22 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 	}
 
 	public void onLink() {
-		IBlockState state = world.getBlockState(linkedSignalPosition);
-		Block block = state.getBlock();
-		if (!(block instanceof SignalBlock)) {
-			if(!world.isRemote)
-				unlink();
-			return;
-		}
+		loadChunkAndGetTile((sigtile, ch) -> {
+			Signal b = Signal.SIGNALLIST.get(sigtile.getBlockID());
 
-		SignalBlock b = (SignalBlock) block;
-
-		HashMap<String, Integer> supportedSignaleStates = new HashMap<>();
-		((IExtendedBlockState) b.getExtendedState(state, world, linkedSignalPosition)).getUnlistedProperties()
-				.forEach((prop, opt) -> opt.ifPresent(x -> {
-					if (prop instanceof SEProperty && ((SEProperty<?>) prop).isChangabelAtStage(ChangeableStage.APISTAGE))
+			HashMap<String, Integer> supportedSignaleStates = new HashMap<>();
+			sigtile.accumulate((bs, prop, obj) -> {
+				if (prop instanceof SEProperty && obj != null) {
+					SEProperty<?> p = ((SEProperty<?>) prop);
+					if (p.isChangabelAtStage(ChangeableStage.APISTAGE)
+							|| p.isChangabelAtStage(ChangeableStage.APISTAGE_NONE_CONFIG))
 						supportedSignaleStates.put(prop.getName(), b.getIDFromProperty(prop));
-				}));
-		listOfSupportedIndicies = supportedSignaleStates.values().stream().mapToInt(Integer::intValue).toArray();
-		tableOfSupportedSignalTypes = supportedSignaleStates;
+				}
+				return null;
+			}, null);
+			listOfSupportedIndicies = supportedSignaleStates.values().stream().mapToInt(Integer::intValue).toArray();
+			tableOfSupportedSignalTypes = supportedSignaleStates;
+		});
 	}
 
 	@Override
@@ -134,12 +136,48 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 		return new Object[] { hasLinkImpl() };
 	}
 
+	public boolean loadChunkAndGetTile(BiConsumer<SignalTileEnity, Chunk> consumer) {
+		TileEntity entity = null;
+		if(linkedSignalPosition == null)
+			return false;
+		Chunk ch = world.getChunkFromBlockCoords(linkedSignalPosition);
+		boolean flag = !ch.isLoaded();
+		if (flag) {
+			if (world.isRemote) {
+				ChunkProviderClient client = (ChunkProviderClient) world.getChunkProvider();
+				ch = client.loadChunk(ch.x, ch.z);
+			} else {
+				ChunkProviderServer server = (ChunkProviderServer) world.getChunkProvider();
+				ch = server.loadChunk(ch.x, ch.z);
+			}
+		}
+		if (ch == null)
+			return false;
+		entity = ch.getTileEntity(linkedSignalPosition, EnumCreateEntityType.IMMEDIATE);
+		boolean flag2 = entity instanceof SignalTileEnity && ((SignalTileEnity)entity).getBlockID() != -1;
+		if (flag2) {
+			consumer.accept((SignalTileEnity) entity, ch);
+		}
+
+		if (flag) {
+			if (world.isRemote) {
+				ChunkProviderClient client = (ChunkProviderClient) world.getChunkProvider();
+				client.unloadChunk(ch.x, ch.z);
+			} else {
+				ChunkProviderServer server = (ChunkProviderServer) world.getChunkProvider();
+				server.queueUnload(ch);
+			}
+		}
+		return flag2;
+	}
+
 	public boolean hasLinkImpl() {
 		if (linkedSignalPosition == null)
 			return false;
-		if (world.getBlockState(linkedSignalPosition).getBlock() instanceof SignalBlock)
+		if (loadChunkAndGetTile((x,y) -> {}))
 			return true;
-		unlink();
+		if (!world.isRemote)
+			unlink();
 		return false;
 	}
 
@@ -156,8 +194,6 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 	}
 
 	public int[] getSupportedSignalTypesImpl() {
-		if (!hasLinkImpl())
-			return new int[] {};
 		return listOfSupportedIndicies;
 	}
 
@@ -176,14 +212,15 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 
 	@SuppressWarnings({ "rawtypes", "unchecked" })
 	public boolean changeSignalImpl(int type, int newSignal) {
-		if (!hasLinkImpl() || !find(getSupportedSignalTypesImpl(), type))
+		if (!find(getSupportedSignalTypesImpl(), type))
 			return false;
-		SignalTileEnity tile = (SignalTileEnity) world.getTileEntity(linkedSignalPosition);
-		IBlockState blockstate = world.getBlockState(linkedSignalPosition);
-		SignalBlock block = (SignalBlock) blockstate.getBlock();
-		SEProperty prop = SEProperty.cst(block.getPropertyFromID(type));
-		tile.setProperty(prop, prop.getObjFromID(newSignal));
-		world.markAndNotifyBlock(linkedSignalPosition, null, blockstate, blockstate, 3);
+		loadChunkAndGetTile((tile, chunk) -> {
+			IBlockState state = chunk.getBlockState(linkedSignalPosition);
+			Signal block = (Signal) state.getBlock();
+			SEProperty prop = SEProperty.cst(block.getPropertyFromID(type));
+			tile.setProperty(prop, prop.getObjFromID(newSignal));
+			world.markAndNotifyBlock(linkedSignalPosition, chunk, state, state, 3);
+		});
 		return true;
 	}
 
@@ -193,8 +230,12 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 		return new Object[] { getSignalTypeImpl() };
 	}
 
+	private String signalTypeCache = null;
+
 	public String getSignalTypeImpl() {
-		return ((SignalBlock) world.getBlockState(linkedSignalPosition).getBlock()).getSignalTypeName();
+		if (signalTypeCache == null)
+			loadChunkAndGetTile((tile, ch) -> signalTypeCache = Signal.SIGNALLIST.get(tile.getBlockID()).getSignalTypeName());
+		return signalTypeCache;
 	}
 
 	@Callback
@@ -205,18 +246,21 @@ public class SignalControllerTileEntity extends TileEntity implements SimpleComp
 
 	@SuppressWarnings("rawtypes")
 	public int getSignalStateImpl(int type) {
-		if (!hasLinkImpl() || !find(getSupportedSignalTypesImpl(), type))
+		if (!find(getSupportedSignalTypesImpl(), type))
 			return -1;
-		SignalTileEnity tile = (SignalTileEnity) world.getTileEntity(linkedSignalPosition);
-		IBlockState blockstate = world.getBlockState(linkedSignalPosition);
-		SignalBlock block = (SignalBlock) blockstate.getBlock();
+		AtomicReference<SignalTileEnity> entity = new AtomicReference<SignalTileEnity>();
+		loadChunkAndGetTile((sig, ch) -> entity.set(sig));
+		SignalTileEnity tile = entity.get();
+		if(tile == null)
+			return -1;
+		Signal block = (Signal) Signal.SIGNALLIST.get(tile.getBlockID());
 		SEProperty prop = SEProperty.cst(block.getPropertyFromID(type));
 		java.util.Optional bool = tile.getProperty(prop);
 		if (bool.isPresent())
 			return SEProperty.getIDFromObj(bool.get());
 		return -1;
 	}
-	
+
 	public BlockPos getLinkedPosition() {
 		return linkedSignalPosition;
 	}